fix(juno): handle negative gas_wanted from failed ante handler transactions
Some cosmos-evm chains return gas_wanted: -1 for transactions that fail during
the ante handler. Changed TxResponse fields from uint64 to int64 to prevent
JSON unmarshalling errors, and guarded negative GasUsed in the gas sum.

+++ packages/juno/node/local/utils.go
@@ -59,9 +59,9 @@ func NewTxResponseFromSdkTxResponse(txResponse *sdk.TxResponse, tx *types.Tx) *t
 	return &types.TxResponse{
 		TxResponse: txResponse,
 		Tx:         tx,
-		Height:     uint64(txResponse.Height),
-		GasWanted:  uint64(txResponse.GasWanted),
-		GasUsed:    uint64(txResponse.GasUsed),
+		Height:    txResponse.Height,
+		GasWanted: txResponse.GasWanted,
+		GasUsed:   txResponse.GasUsed,
 	}
 }
